dive-time widget: fix incorrect use of timestamp_t
I did a global search-and-replace to make all "time_t" users use the
internal subsurface 64-bit "timestamp_t" type instead, but we have one
case that still uses the system time functions: the use of "localtime()"
in the dive_time_widget().

Everywhere else we always just use UTC for all our time handling, and we
don't really ever care about the local timezone etc.  However, for the
dive time widget, we initialize the calendar widget to the current time,
which obviously does want to take the local timezone into account, so
there we end up using the whole system time handling code.

So that one should continue to use time_t, even if it might have the
year-2038 problem.  We also don't care about the fact that it's not
thread-safe, since this is just initializing the widget which definitely
doesn't happen threaded.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>