Fix incorrect localization of dialog name
The existing code passed a localized copy of a text buffer initialized by
a static string to a function that ended up wanting to modify the buffer.
Unsurprisingly, that doesn't work. This commit restructures the code so
that we initialize the buffer at run time with a localized version of the
default string and then just pass the buffer around.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>