Add Makefile target to update po files
This is a bit of a hack to make my life easier.

make update-po-files

will extract the translation strings and merge them with the existing
translations - for all existing translations.

For good measure this commit includes the latest update of the po files
(but no new translations should be needed).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>