Changed the term used for one of the default weight systems
Originally I used the colloquial term "bar" for the weight bar that you
can attach to a backplate (usually when diving twins in a tec
environment). This of course causes an odditity for translations as this
word "bar" is a homonym for the weight system and the pressure unit -
which throws off translations.

Instead of switching to a context-based translation I instead went with a
better term: "backplate weight". This of course now needs to be
translated, so I updated the .po files (and added German translations for
the two flavors of German).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>