Hopefully the last update to the po files before the release
This one breaks with my preference to separate generated changes from
content changes - I wanted to get the new comment next to the
translator-credits text that I added to every .po file. This way the
people who worked on these translations at least get shown in the About
box. But a simple grep on the diff will show you that this is indeed the
only set of changes that I made.

git diff HEAD^ | grep ^+ | grep -v -e^+# -e^+++ -ePOT-Creation

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>