Display air temperature in the info notebook page
So far we don't parse air temperature data via libdivecomputer. Nor are we
set up to allow the user to manually enter it. We can parse it when
downloading from a Uemis Zurich, though.

This feature was suggested via trac.hohndel.org; this commit implements
only part of what is requested there, so I simply reference the ticket
here without closing it.

References ticket #7

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>